+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 - **IntelliJ Platform Plugin Template** – Upgraded IntelliJ Platform Plugin Template to [`v2.2.0`](https://github.com/JetBrains/intellij-platform-plugin-template/releases/tag/v2.2.0).<br>
   ⚠️ `master` branch has to be renamed to `main` to follow the same GitHub actions rules and help futur upgrades.
 
+### Fixed
+
+- **Intention definition** - Add missing `JSON` language to JCV replacement intention configuration.
+- **API usage** - Remove "override-only" api usage and deprecated ones.
+- **File type** - Replaced object by class
+
 ## [3.0.3] - 2025-07-24
 
 ### Changed